Move specs around for business logic.
