+++ core/Gemfile
@@ -1,0 +1,18 @@
+source "http://rubygems.org"
+
+gemspec
+
+git 'git://github.com/rails/rails.git' do
+  gem 'activemodel'
+end
+
+group :development, :test do
+  gem 'rspec'
+  gem 'simplecov'
+end
+
+group :development do
+  gem 'yard', "~> 0.8.4.1"
+  gem 'rdoc'
+  gem 'kramdown'
+end
