+++ core/entities/menu/menu_entry.rb
