+++ core/entities/menu/menuable.rb
