+++ core/entities/menu/path.rb
