+++ core/entities/view/viewable.rb
