+++ core/entities/view/viewable_presenter.rb
