+++ core/spec/entities/menu/menu_entry_spec.rb
@@ -1,6 +1,5 @@
 require 'spec_helper'
-require 'abc/menu/menu_entry'
-require 'ostruct'
+require 'menu/menu_entry'
 
 module Abc
   describe MenuEntry do
