--- spec/spec_helper.rb
@@ -1,4 +1,0 @@
-require 'simplecov'
-SimpleCov.start do
-  add_filter "/spec/"
-end
